all for redeem operations

+++ aaas-finance-trade/aaas-finance-trade-pay/src/main/java/com/shellshellfish/aaas/finance/trade/pay/service/FundTradeApiService.java
@@ -13,7 +13,8 @@ public interface FundTradeApiService {
 
     BuyFundResult buyFund(String userUuid, String tradeAcco, BigDecimal applySum, String outsideOrderNo, String fundCode) throws Exception;
 
-    SellFundResult sellFund(String userUuid, Integer sellNum, String outsideOrderNo, String tradeAcco, String fundCode) throws Exception;
+    SellFundResult sellFund(String userUuid, BigDecimal sellNum, String outsideOrderNo, String tradeAcco,
+        String fundCode) throws Exception;
 
     FundConvertResult fundConvert(String userUuid, BigDecimal applyShare, String outsideOrderNo,
         String tradeAcco,String fundCode, String targetFundCode ) throws Exception;
